Implement remote querying of all flavors
Implements a Fog request handler and response
parser to enable retrieving a list of
all flavors.

This functionality is specific to the ec2 api
implementation of nova and is not currently 
supported by AWS. Nova, unlike AWS allows arbitrary
flavor names.

The 'describe_instance_types' api call is only
supported by the admin controller of the nova
ec2 api. This contoller needs to be explicitly
enabled by using the --allow_ec2_admin_api flag,
and has a separate endpoint than the cloud controller
(which services the rest of the ec2 compatibility
APIs)