Add param to allow flavors to be queried remotely
This commit adds a parameter called query_endpoint
to both the flavors.all as well as the flavors.get
methods.

This parameter defaults to false (so as not to break
backwards compatibility).

Setting the parameter to true causes the list of
flavors to be queried using the desribe_instance_types
request. (as opposed to the hardcoded list of
aws flavors)